src/mandelbroth.pth: Fix division remainder error
Before this, if the image size was not multiple of MAX_THREADS, the last
few pixels were not written to the image. The problem was that

(image_size/MAX_THREADS) * MAX_THREADS != image_size

Now the threads called by the first for loop are in charge of the
remainder of the division.